fix folder hierarchy
