+++ packages/components/modules/payments/web/ProductGallery/Product/BuyButton/index.tsx
